backend/app/core/metrics.py
from starlette.middleware.base import BaseHTTPMiddleware
from fastapi import Request

# Simple Prometheus Middleware Stub
# If we had full prom-client, we'd use it. For now, we mock or simpler version.
# Main.py imports: PrometheusMiddleware, metrics_endpoint

class PrometheusMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next):
        # Measure time?
        response = await call_next(request)
        return response

def metrics_endpoint(request: Request):
    from fastapi.responses import PlainTextResponse
    # Return empty metrics or basic stats
    return PlainTextResponse("# HELP http_requests_total Total number of HTTP requests\n# TYPE http_requests_total counter\n")
