backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.30
alembic>=1.13.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
cryptography==42.0.2
python-multipart==0.0.9
email-validator==2.1.0.post1

networkx==3.2.1
requests==2.31.0
google-generativeai==0.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0
